exercises: 11 - 16 + bonus task

+++ controllers/blogs.js
@@ -1,18 +1,54 @@
 const blogsRouter = require('express').Router()
-const { Blog } = require('../models/index')
+const { Op } = require('sequelize')
+const { Blog, User } = require('../models/index')
+const { tokenExtractor } = require('../util/middleware')
 
-blogsRouter.get('/', async (_request, response, next) => {
+blogsRouter.get('/', async (request, response, next) => {
   try {
-    const blogs = await Blog.findAll()
+    const where = {}
+
+    if(request.query.search) {
+      where[Op.or] = [
+          {
+            title: {
+              [Op.iLike]: `%${request.query.search}%`
+            }
+          },
+          {
+            author: {
+              [Op.iLike]: `%${request.query.search}%`
+            }
+          }
+        ]
+      }
+    
+
+    const blogs = await Blog.findAll({
+      attributes: {
+        exclude: ['userId']
+      },
+      include: {
+        model: User,
+        attributes: ['name']
+      },
+      where,
+      order: [
+        ['likes', 'DESC']
+      ]
+    })
+
     response.json(blogs)
   } catch (error) {
     next(error)
   }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', tokenExtractor, async (request, response, next) => {
   try {
-    const blog = await Blog.create(request.body)
+    const date = new Date().toString().substring(16, 34).replace(' GMT', '')
+    const user = await User.findByPk(request.decodedToken.id)
+
+    const blog = await Blog.create({...request.body, userId: user.id, date})
     response.json(blog)
   } catch (error) {
     next(error)
@@ -30,14 +66,25 @@ blogsRouter.put('/:id', async (request, response, next) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', tokenExtractor,  async (request, response, next) => {
   try {
-    await Blog.destroy({
-      where: {
-        id: request.params.id
+    const user = await User.findByPk(request.decodedToken.id, {
+      include: {
+        model: Blog,
+        attributes: ['userId']
       }
     })
-    response.status(204).end()
+
+    if(user.blogs.some(blog => blog.userId === user.id)) {
+      await Blog.destroy({
+        where: {
+          id: request.params.id
+        }
+      })
+      response.status(204).end()
+    } else {
+      response.status(404).send({ error: 'Unauthorized token' })
+    }
   } catch (error) {
     next(error)
   }
